+++ src/bot.js
@@ -4,77 +4,25 @@
  * Licensed under Lesser General Public License v2.1 (LGPl-2.1 - https://opensource.org/licenses/lgpl-2.1.php)
  */
 const fs = require("fs");
-const Discord = require("discord.js");
+const WelcomeBot = require("./WelcomeBot");
 const dotenv = require("dotenv").config();
 
-const client = new Discord.Client();
+const client = new WelcomeBot();
 
-client.commands = new Discord.Collection();
-client.disabled = new Discord.Collection();
-client.cooldowns = new Discord.Collection();
-client.defaultPrefix = "w/";
-client.botVersion = "1.4.0";
-client.changelog = [
-    "**Bug Fixes**",
-    "Fix bug `msg not working` (Issue: [#19]) ([#20])",
-    "Fix goodBye",
-    "**Changes**",
-    "Add new subcommands for `channel` command - `setMod`, `getMod`, `resetMod` ([#20])",
-    "Add new command version ([#20])",
-    "Moderation logger also! ([#20])",
-    "Add `restart` command for owner ([#20])",
-];
 process.env.userAgent = "Discord Welcome-Bot " + client.botVersion;
 process.env.ownerIDs = [
-    "815204465937481749" /*PuneetGopinath#6398*/,
+    "815204465937481749" /*PuneetGopinath#6300*/,
     "693754859014324295" /*abhijoshi2k#6842*/,
 ];
 process.on("exit", (code) => {
     client.destroy();
 });
 
-const commandFolder = __dirname + "/commands";
-const commandFolders = fs.readdirSync(commandFolder);
-let defaultOpts = {
-    bot_perms: ["VIEW_CHANNEL", "SEND_MESSAGES", "READ_MESSAGE_HISTORY"],
-    args: false,
-    catchError: true,
-    disabled: false,
-    cooldown: 3,
-    ownerOnly: false,
-};
-
-for (const folder of commandFolders) {
-    const commandFiles = fs
-        .readdirSync(`${commandFolder}/${folder}`)
-        .filter((file) => file.endsWith(".js"));
-    for (const file of commandFiles) {
-        let module = file.replace(".js", "");
-        let command = require(`${commandFolder}/${folder}/${module}`);
-        if (command.bot_perms) {
-            command.bot_perms = [
-                ...defaultOpts.bot_perms,
-                ...command.bot_perms,
-            ];
-        }
-        command = {
-            ...defaultOpts,
-            ...command,
-        };
-        if (!command.disabled) {
-            client.commands.set(command.name, command);
-        } else {
-            client.disabled.set(command.name, command);
-        }
-    }
-}
-
 const presence = require("./functions/presence");
 const greetUser = require("./functions/greetUser");
 const sayGoodBye = require("./functions/sayGoodBye");
 const serverCount = require("./functions/serverCount");
 const execute = require("./functions/execute");
-const uptime = require("./functions/uptime");
 
 require("./db/connection");
 const addGuild = require("./db/functions/guild/addGuild");
@@ -84,7 +32,9 @@ const dbAuditor = require("./db/functions/dbAuditor");
 
 client.on("ready", () => {
     // We logged in
-    console.log(`Logged in as ${client.user.tag}!`);
+    console.log(
+        `${client.user.tag}, ready to serve ${client.users.cache.size} users in ${client.guilds.cache.size} servers.`
+    );
     process.env.BOT_ID = client.user.id;
     presence(client);
     if (process.env.NODE_ENV === "production") serverCount(client);
@@ -94,9 +44,6 @@ client.on("ready", () => {
     // Update server count every 25 minutes if environment is in PRODUCTION
     if (process.env.NODE_ENV === "production")
         setInterval(() => serverCount(client), 25 * 60 * 1000);
-    uptime(client);
-    //Log uptime every 15 minutes
-    setInterval(() => uptime(client), 15 * 60 * 1000);
     //Run dbAuditor every 3 hours
     setInterval(() => {
         dbAuditor(client);
@@ -120,6 +67,9 @@ client.on("guildMemberRemove", (member) => {
 client.on("guildCreate", (guild) => {
     //Bot has been invited to a new guild
     addGuild(guild.id);
+    guild.channels.cache
+        .find((ch) => ch.id === guild.systemChannelID)
+        .send("Thank you for choosing this bot! To get started, type `w/help`");
 });
 
 //https://discord.js.org/#/docs/main/v12/class/Client?scrollTo=e-guildDelete
@@ -134,7 +84,7 @@ client.on("message", async function (message) {
     if (message.guild && message.channel.type !== "dm") {
         guildDB = await getGuild(message.guild.id);
     } else {
-        guildDB = { prefix: "w/" };
+        guildDB = { prefix: client.defaultPrefix };
     }
 
     if (message.mentions.has(client.user)) {
@@ -161,7 +111,7 @@ client.on("message", async function (message) {
         }
     }
     //Whatever happens, if the message starts with prefix, the bot should execute it.
-    execute(message, client, guildDB);
+    execute(message, guildDB);
 });
 
 // Login