+++ src/functions/presence.js
@@ -3,16 +3,21 @@
  * Copyright (c) 2021 The Welcome-Bot Team and Contributors
  * Licensed under Lesser General Public License v2.1 (LGPl-2.1 - https://opensource.org/licenses/lgpl-2.1.php)
  */
-const presence = function (client) {
+module.exports = function (client) {
     const servers = client.guilds.cache.size;
+    const chans = client.channels.cache.size;
+    const users = client.users.cache.size;
     console.log(`Updating presence. Servers: ${servers}`);
     client.user
         .setPresence({
             activity: {
-                name: `w/help | ${servers} server${servers > 1 ? "s" : ""}`,
+                name: `w/help | ${servers} server${
+                    servers > 1 ? "s" : ""
+                } | handling ${chans} channel${
+                    chans > 1 ? "s" : ""
+                } | giving hand to ${users} user${users > 1 ? "s" : ""}`,
                 type: "WATCHING",
             },
         })
         .catch((error) => console.error(error));
 };
-module.exports = presence;